// frontend/projects/partner/src/app/api/model/partner-group-search-dto.ts
/**
 * BDK-CoreFinance API Gateway
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Partner group search result containing comprehensive partner and group information
 */
export interface PartnerGroupSearchDto {
  /**
   * Partner number - primary partner identifier
   */
  partnerNumber?: number;
  /**
   * Group number - primary group identifier
   */
  groupNumber?: number;
  /**
   * Entity type (partner vs. group)
   */
  type?: string;
  /**
   * Alpha code identifier / short name
   */
  alphaCode?: string;
  /**
   * Name part 1
   */
  name1?: string;
  /**
   * Name part 2
   */
  name2?: string;
  /**
   * Name part 3
   */
  name3?: string;
  /**
   * First name of the partner
   */
  firstname?: string;
  /**
   * Street address
   */
  street?: string;
  /**
   * House number of the partner\'s address
   */
  houseNumber?: string;
  /**
   * City/town of the partner\'s address
   */
  city?: string;
  /**
   * Postal code of the partner\'s address
   */
  postalCode?: string;
  /**
   * Group type (internal vs. external)
   */
  groupType?: string;
}